+++ Example/Tests/Doubles/BitmovinPlayerTestDouble.swift
@@ -70,6 +70,7 @@ class BitmovinPlayerTestDouble: BitmovinPlayerStub, TestDoubleDataSource {
         onPlayerError(
             PlayerErrorEvent(
                 code: PlayerError.Code.networkGeneral,
+                errorCode: 1_200,
                 message: "Test player error",
                 data: nil
             ),
@@ -81,7 +82,15 @@ class BitmovinPlayerTestDouble: BitmovinPlayerStub, TestDoubleDataSource {
         guard let onSourceError = fakeListener?.onSourceError else {
             return
         }
-        onSourceError(SourceErrorEvent(code: SourceError.Code.general, message: "Test source error", data: nil), self)
+        onSourceError(
+            SourceErrorEvent(
+                code: SourceError.Code.general,
+                errorCode: 2_001,
+                message: "Test source error",
+                data: nil
+            ),
+            self
+        )
     }
 
     func fakeAdStartedEvent(position: String? = "pre") {